Added the functions and basic interface for motors

+++ src/main/java/com/fireteam322/frc/robot/interfaces/MotorInterface.java
@@ -3,7 +3,7 @@ package com.fireteam322.frc.robot.interfaces;
 public interface MotorInterface {
     public void run();
 
-    public void run(int speed);
+    public void run(double speed);
 
     public void stop();
 }
